backend/pyproject.toml
[project]
name = "zoom-ai-assistant"
version = "0.1.0"
description = "Real-Time AI Executive Assistant for Zoom Meetings"
authors = [
    {name = "AI Assistant", email = "davidgundaker9@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "openai>=1.3.0",
    "openai-whisper>=20231117",
    "elevenlabs>=0.2.26",
    "pyaudio>=0.2.11",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "asyncio-mqtt>=0.16.1",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.1",
    "pydantic>=2.5.0",
    "httpx>=0.25.2",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sounddevice>=0.4.6",
    "webrtcvad-wheels>=2.0.10",
    "librosa>=0.10.1",
    "pydantic-settings>=2.9.1",
    "requests>=2.31.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]
